+++ next.config.mjs
@@ -1,5 +1,17 @@
 const nextConfig = {
   reactStrictMode: true,
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "commons.wikimedia.org"
+      },
+      {
+        protocol: "https",
+        hostname: "upload.wikimedia.org"
+      }
+    ]
+  }
 };
 
 export default nextConfig;